Version 7.5.0.43
Reports against players no longer expire at round end. 500 reports are kept in memory before being removed. Reports are updated in AdKats if they are updated in the BFACP. An 'active' report is now a report that has not been confirmed, denied, ignored, or expired. A report is expired when a player leaves their current gaming session with a report against them being active.